Gossip v1: Inefficient, converges slowly, each neighbour is communicated with same probability, does not converge when nodes have less neighbours, assumes nodes are uniformly distributed on the field.

+++ Gossip.h
@@ -28,7 +28,7 @@ class Gossip: public VirtualApplication {
   int packetsSent;
   simtime_t neighbourCheckInterval, gossipInterval;
   vector<int> peers;
-  int gossipMsg;
+  double gossipMsg;
   bool isBusy;
   short currentPeerIndex;
   int expectedSeq;
